feat: initial extraction to a dedicated library
