+++ .gitignore
@@ -1,2 +1,3 @@
 .idea
 TODO.md
+*.egg-info
