+++ tests/test_file_outputs.py
@@ -1,6 +1,5 @@
 import pytest
-
-from grabit import sanitize_filename
+from grabit_lib.core import sanitize_filename
 
 
 @pytest.mark.parametrize(
