+++ tests/test_markdown_conversion.py
@@ -1,4 +1,4 @@
-from grabit import convert_to_markdown
+from grabit_lib.core import convert_to_markdown
 
 
 def test_converts_bold_and_italics_with_different_markup():
